@@ -19,6 +19,7 @@
 	"funding": "https://github.com/sponsors/fregante",
 	"license": "MIT",
 	"author": "Federico Brigante <me@fregante.com> (https://fregante.com)",
+	"type": "module",
 	"exports": "./index.js",
 	"main": "./index.js",
 	"types": "./index.d.ts",
@@ -30,6 +31,7 @@
 		"test": "xo && tsd && ava"
 	},
 	"xo": {
+		"parser": "@typescript-eslint/parser",
 		"rules": {
 			"unicorn/better-regex": "off",
 			"unicorn/prefer-module": "off",
@@ -40,12 +42,12 @@
 		"github-reserved-names": "^2.0.5"
 	},
 	"devDependencies": {
-		"@sindresorhus/tsconfig": "^4.0.0",
-		"ava": "^5.3.1",
-		"tsd": "^0.29.0",
-		"xo": "^0.56.0"
+		"@sindresorhus/tsconfig": "^5.0.0",
+		"ava": "^6.1.3",
+		"tsd": "^0.31.1",
+		"xo": "^0.58.0"
 	},
 	"engines": {
-		"node": ">=18"
+		"node": ">=20.10"
 	}
 }
