@@ -7,17 +7,14 @@ Shortens a GitHub URL string.
 @example https://github.com/nodejs/node/tree/v0.12/doc becomes nodejs/node@<code>v0.12</code>
 
 */
-declare function shortenRepoUrl(url: string, currentUrl?: string): string;
+export default function shortenRepoUrl(url: string, currentUrl?: string): string;
 
-declare namespace shortenRepoUrl {
-	/**
-	Shortens a GitHub URL in a DOM anchor.
+/**
+Shortens a GitHub URL in a DOM anchor if the link label is not customized (i.e. if the `href` matches the `textContent`)
 
-	@param anchor An HTMLAnchorElement
-	@param url The GitHub URL to shorten.
-	@example https://github.com/nodejs/node/tree/v0.12/doc becomes nodejs/node@<code>v0.12</code>
+@param anchor An HTMLAnchorElement
+@param url The GitHub URL to shorten.
+@example<a href="https://github.com">https://github.com</a> becomes <a href="https://github.com">github.com</a>
 
-	*/
-	const applyToLink: (anchor: HTMLAnchorElement, url?: string) => void;
-}
-export = shortenRepoUrl;
+*/
+export function applyToLink(anchor: HTMLAnchorElement, url?: string): void;
